% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_df.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape a string for safe use in regular expressions}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{Character vector to escape.}
}
\value{
Character vector with metacharacters escaped.
}
\description{
Escape a string for safe use in regular expressions
}
