% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bipnet.R
\name{GeomBipnetInteractionBinary}
\alias{GeomBipnetInteractionBinary}
\title{Geom for bipartite interaction segments (binary mode)}
\usage{
GeomBipnetInteractionBinary
}
\description{
\code{GeomBipnetInteractionBinary} draws one segment per row-column interaction.
It is used when \code{interaction_type = "binary"} in
\code{\link[=geom_bipnet_interaction]{geom_bipnet_interaction()}}.
}
\keyword{internal}
