% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipartite-package.R
\docType{package}
\name{ggbipartite-package}
\alias{ggbipartite}
\alias{ggbipartite-package}
\title{ggbipartite: Visualize Bipartite Networks with ggplot2}
\description{
Provides tools to visualize bipartite interaction matrices on of ggplot2. It includes stat/geom layers to draw row and column boxes and interaction polygons, coordinate constructors for custom workflows, and helpers to align bipartite layouts with phylogenetic trees.
}
\author{
\strong{Maintainer}: Hidehito Okada \email{hidefungi@gmail.com} (\href{https://orcid.org/0000-0002-3210-2736}{ORCID})

Authors:
\itemize{
  \item Hidehito Okada \email{hidefungi@gmail.com} (\href{https://orcid.org/0000-0002-3210-2736}{ORCID})
}

}
\keyword{internal}
