% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_df.R
\name{prefix_cols}
\alias{prefix_cols}
\title{Prefix all column names with "<prefix><sep>"}
\usage{
prefix_cols(df, prefix, sep = "__")
}
\arguments{
\item{df}{A data frame / tibble.}

\item{prefix}{Scalar character to prepend.}

\item{sep}{Separator between prefix and original name (default: "__").}
}
\value{
A tibble with renamed columns.
}
\description{
Prefix all column names with "<prefix><sep>"
}
