% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bipnet.R
\docType{data}
\name{GeomBipnetInteractionBinary}
\alias{GeomBipnetInteractionBinary}
\title{Geom for bipartite interaction segments (binary mode)}
\format{
An object of class \code{GeomBipnetInteractionBinary} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomBipnetInteractionBinary
}
\description{
`GeomBipnetInteractionBinary` draws one segment per row-column interaction.
It is used when `interaction_type = "binary"` in
[geom_bipnet_interaction()].
}
\keyword{internal}
