% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbipartite-package.R
\docType{package}
\name{ggbipartite-package}
\alias{ggbipartite}
\alias{ggbipartite-package}
\title{ggbipartite: Visualize Bipartite Networks with ggplot2}
\description{
Provides tools to visualize bipartite interaction matrices with 'ggplot2'. It computes reusable layouts for abundance ribbons or binary links, offers layers and a high-level plot composer, and aligns either network partition with a phylogenetic tree.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Hide-Fun/ggbipartite}
  \item Report bugs at \url{https://github.com/Hide-Fun/ggbipartite/issues}
}

}
\author{
\strong{Maintainer}: Hidehito Okada \email{hidefungi@gmail.com} (\href{https://orcid.org/0000-0002-3210-2736}{ORCID})

Authors:
\itemize{
  \item Hidehito Okada \email{hidefungi@gmail.com} (\href{https://orcid.org/0000-0002-3210-2736}{ORCID})
}

}
\keyword{internal}
